finished styling of profile page.
